move tests/test_visualization_new_structure.py
